Finalize Sanaei updateClient inbound-id int fix and diagnostics
Agent-Logs-Url: https://github.com/hforouzi/amoobot/sessions/39a704d0-3f3b-43aa-8b56-49567d6fc8ba

+++ src/Provisioning/Application/ServiceRenewalService.php
@@ -75,8 +75,8 @@ final class ServiceRenewalService
                 trafficLimitGb: $newTrafficLimitGb,
                 expiresAt: $newExpiresAt,
                 unlimitedDuration: $unlimitedDuration,
-                serviceId: (int) ($service->getId() ?? 0),
-                orderId: (int) ($order->getId() ?? 0),
+                serviceId: $service->getId() ?? 0,
+                orderId: $order->getId() ?? 0,
             ),
             $service->getPanel()
         );
